Upstream Sultan Updates
Revert "tmpfs: clear S_ISGID when setting posix ACLs"

This reverts commit b0f454b1a49ec08417071fc9aa1ff9705f2974d4.

Change-Id: I1ceeabd2a3a0c452c20f37122595376845307dfb

power: bq27541: Fix hardware config init failing on some devices

The init delay was not actually 300 ms, and upon failing to configure the
device, no attempt to retry was made.

Increase the init delay and retry in the event of failure to ensure that
init is always successful on all devices.

Signed-off-by: Sultanxda <sultanxda@gmail.com>